Ironing out some kinks in my function code gen...
* The symbol table should have a field that
communicates what kind of symbol is stored in a
given entry (a variable or a function). We field
should be used when generating code for variables.
* Currently, we are unable to do recursion because
the function value is stored in the symbol table
AFTER generating the function body. A way to get
around this limitation is to do the following:
  1. When generating code for `LambdaNode`, go up
  to the point where we have generated created a
  signature.
  2. Check if this lambda is anonymous or not.
  If it is not anonymous, get its identifier,
  store the signature in the symbol table entry
  corresponding to the identifier, and continue
  on to generating the lambda body.
  If it is anonymous, continue on to generating
  the lambda body.
* This will allow us to simply call GenerateCode
on the lambda child of LambdaInvocationNode
(assuming `GenerateCode(IdentifierNode&)` is
changed as well), which would also allow us to
call anonymous lambdas!